Added Harris corner detector
When a user selects their points, a small function will run in the
background to look around the immediate area to find a “better” point
for them. This way we can eliminate some of the instances of drifting.

Will need to explore a better search/threshold function in the future.